--- templates/forgot_password/request.html.twig
@@ -1,9 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block body %}
-<h1>Réinitialisation du mot de passe</h1>
-{{ form_start(form) }}
-{{ form_row(form.email) }}
-<button type="submit">Envoyer</button>
-{{ form_end(form) }}
-{% endblock %}
