+++ templates/security/forgot_password.html.twig
@@ -1,0 +1,27 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Mot de passe oublié{% endblock %}
+
+{% block body %}
+<section class="container my-5">
+    <h1 class="text-center mb-4">Mot de passe oublié</h1>
+    <div class="row justify-content-center">
+        <div class="col-md-6 col-lg-4">
+            <div class="card shadow-sm">
+                <div class="card-body">
+                    {{ form_start(form, { attr: { class: 'needs-validation', novalidate: 'novalidate' } }) }}
+                        <div class="mb-3">
+                            {{ form_label(form.email, 'Adresse e-mail', { label_attr: { class: 'form-label' } }) }}
+                            {{ form_widget(form.email, { attr: { class: 'form-control', placeholder: 'Votre adresse e-mail' } }) }}
+                            {{ form_errors(form.email) }}
+                        </div>
+                        <div class="d-grid">
+                            <button type="submit" class="btn btn-primary">Envoyer</button>
+                        </div>
+                    {{ form_end(form) }}
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
+{% endblock %}
