+++ templates/security/reset_password.html.twig
@@ -1,0 +1,27 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Réinitialisation du mot de passe{% endblock %}
+
+{% block body %}
+<section class="container my-5">
+    <h1 class="text-center mb-4">Réinitialisation du mot de passe</h1>
+    <div class="row justify-content-center">
+        <div class="col-md-6 col-lg-4">
+            <div class="card shadow-sm">
+                <div class="card-body">
+                    {{ form_start(form, { attr: { class: 'needs-validation', novalidate: 'novalidate' } }) }}
+                        <div class="mb-3">
+                            {{ form_label(form.plainPassword, 'Nouveau mot de passe', { label_attr: { class: 'form-label' } }) }}
+                            {{ form_widget(form.plainPassword, { attr: { class: 'form-control', placeholder: 'Votre nouveau mot de passe' } }) }}
+                            {{ form_errors(form.plainPassword) }}
+                        </div>
+                        <div class="d-grid">
+                            <button type="submit" class="btn btn-primary">Réinitialiser</button>
+                        </div>
+                    {{ form_end(form) }}
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
+{% endblock %}
